requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Werkzeug==3.0.1
google-generativeai==0.8.3
python-dotenv==1.0.0
Pillow==11.0.0
SpeechRecognition==3.10.4
pydub==0.25.1